Extract text ellipsizing and scaling into TextFitter and improve logic
